% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDendroPaint.R
\name{pintaRDendroPaint}
\alias{pintaRDendroPaint}
\title{pintaRDendroPaint}
\usage{
pintaRDendroPaint(ListaSal)
}
\arguments{
\item{ListaSal}{input DendroTree}
}
\description{
Paint in console the DendroTree returned by RDendroPaint with the generic visualization
}
