% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowDendro.R
\name{WriteDendro}
\alias{WriteDendro}
\title{WriteDendro}
\usage{
WriteDendro(valoresMatriz, tablaMatrix, masdocumenWrite = FALSE,
  masvacios = FALSE)
}
\arguments{
\item{valoresMatriz}{Tree Matrix with all grups generatod by, from 2 to limit-1 cutree(hclust_avg, k =2:(length(hclust_avg$order)-1))}

\item{tablaMatrix}{Cross original matrix Document-Term with relation 1 or 0 if term exist or not}

\item{masdocumenWrite}{True for extra info of documents, default False if no estra info}

\item{masvacios}{Paint emptry cross elements groups}
}
\value{
list with non empty element groups
}
\description{
Paint in console the dondrogram and return a list with non empty cross elements group
}
